Separation of SupportTable
SupportTable moved to single class.
Bug with file path on linux fixed.
